move builtin helper to separate code
