+++ services/slam/builtin/builtin.go
@@ -491,28 +491,7 @@ func (slamSvc *builtIn) GetPosition(ctx context.Context, name string) (spatialma
 	componentReference := resp.GetComponentReference()
 	returnedExt := resp.Extra.AsMap()
 
-	return CheckQuaternionFromClientAlgo(pose, componentReference, returnedExt)
-}
-
-// CheckQuaternionFromClientAlgo checks to see if the internal SLAM algorithm sent a quaternion. If it did, return the updated pose
-func CheckQuaternionFromClientAlgo(pose spatialmath.Pose, componentReference string, returnedExt map[string]interface{}) (spatialmath.Pose, string, error) {
-	// check if extra contains a quaternion. If no quaternion is found, throw an error
-	if val, ok := returnedExt["quat"]; ok {
-		q := val.(map[string]interface{})
-
-		valReal, ok1 := q["real"].(float64)
-		valIMag, ok2 := q["imag"].(float64)
-		valJMag, ok3 := q["jmag"].(float64)
-		valKMag, ok4 := q["kmag"].(float64)
-
-		if !ok1 || !ok2 || !ok3 || !ok4 {
-			return nil, "", errors.Errorf("error getting SLAM position: quaternion given, but invalid format detected, %v", q)
-		}
-		actualPose := spatialmath.NewPose(pose.Point(),
-			&spatialmath.Quaternion{Real: valReal, Imag: valIMag, Jmag: valJMag, Kmag: valKMag})
-		return actualPose, componentReference, nil
-	}
-	return nil, "", errors.Errorf("error getting SLAM position: quaternion not given, %v", returnedExt)
+	return checkQuaternionFromClientAlgo(pose, componentReference, returnedExt)
 }
 
 // GetMap forwards the request for map data to the slam library's gRPC service. Once a response is received it is unpacked
